+++ webdev/lib/src/command/configuration.dart
@@ -261,7 +261,7 @@ class Configuration {
 
   bool get requireBuildWebCompilers => _requireBuildWebCompilers ?? true;
 
-  bool get enableExpressionEvaluation => _enableExpressionEvaluation ?? false;
+  bool get enableExpressionEvaluation => _enableExpressionEvaluation ?? true;
 
   bool get verbose => _verbose ?? false;