+++ webdev/lib/src/serve/webdev_server.dart
@@ -8,7 +8,7 @@ import 'dart:async';
 import 'dart:io';
 
 import 'package:build_daemon/data/build_status.dart' as daemon;
-import 'package:devtools_server/devtools_server.dart';
+import 'package:dds/devtools_server.dart';
 import 'package:dwds/data/build_result.dart';
 import 'package:dwds/dwds.dart';
 import 'package:http/http.dart' as http;
@@ -150,8 +150,11 @@ class WebDevServer {
           expressionCompiler: ddcService,
           devtoolsLauncher: shouldServeDevTools
               ? (String hostname) async {
-                  var server = await serveDevTools(
-                      hostname: hostname, enableStdinCommands: false);
+                  var server = await DevToolsServer().serveDevTools(
+                    hostname: hostname,
+                    enableStdinCommands: false,
+                    customDevToolsPath: devToolsPath,
+                  );
                   return DevTools(server.address.host, server.port, server);
                 }
               : null);